Close on no tabs, smoothed animations
I made the animation for when a tab spawns smoother, and also added an animation for when tab closes. the second one causes an issue tho, were if you closed two tabs before the 400ms passed, you could confuse the browser and no tab would be focused. I added a little fix, but its only able to run after those 450ms passed, so thats definetaly smth we have to look into for the future. For now the closing animation will be reduced to 150-200ms to reduce the time frame in which you can glitch it out.